Update and implement Pith::Page
Since the page size cannot be statically set at compile time, we can't
use the Page class as the literal representation of a page.

So instead we use the Process::properties()::pageSize(), and the Page
APIs now work with pointers and size_t's. The Span<Page> code was
awkward anyways.

Signed-off-by: Robert Young <rwy0717@gmail.com>