Implement a WASM dump utility
As a test case for the WASM binary reader, create a new utility:
ab-dump. This will output the human readable s-expression format of the
WASM module. The support is incomplete and messy. At this stage, we're
still not parsing all the instructions.

I think there's still quite a bit to figure out around nesting parsing,
context, and error checking, but for now, we're very close to the WABT
output, which is good enough.

This massive commit also introduces a hacky framework for printing
s-expressions and indenting output to streams.

Signed-off-by: Robert Young <rwy0717@gmail.com>