Clean up the SexprPrinter
Since it looks like the sexpr printer is here to stay (hey, it's
useful), we might as well clean up the APIs and document it all.

The SexpPrinter can still be tweaked by users, I specifically didn't
close off any state, so the sexpr printer doesn't "guarantee" that
things will be nicely formatted.

It's too hard to predict what the result of any kind of raw output will
be, and of course we want to allow raw output, or let users build their
own manipulators.

Signed-off-by: Robert Young <rwy0717@gmail.com>